dentalXray/backend/requirements.txt
fastapi
uvicorn
pydicom
pillow
python-multipart
requests
google-generativeai
